Move AuthorityData from proto to recursor
